requirements.txt
# aiohttp==3.6.2
# APScheduler==3.6.3
# args==0.1.0
# asgiref==3.4.1
# asn1crypto==0.24.0
# async-timeout==3.0.1
# attrs==22.2.0
# Automat==22.10.0
# backports.zoneinfo==0.2.1
# beautifulsoup4==4.12.3
# bs4==0.0.2
# CacheControl==0.12.14
# cachetools==3.1.1
# calc==1.0
# certifi==2019.6.16
# cffi==1.12.3
# cfscrape==2.0.8
# chardet==3.0.4
# charset-normalizer==2.0.12
# click==8.0.4
# colorama==0.4.5
# constantly==15.1.0
# cryptography==37.0.4
# cssselect==1.0.3
# data==0.4
# dataclasses==0.8
# decorator==4.4.0
# Django==3.2.23
# django-cors-headers==3.10.1
# django-environ==0.11.2
# django-filter==21.1
# django-messages==0.6.0
# djangorestframework==3.14.0
# djangorestframework-simplejwt==4.4.0
# et-xmlfile==1.1.0
# filelock==3.4.1
# firebase-admin==5.4.0
# Flask==2.0.3
# Flask-Cors==4.0.0
# funcsigs==1.0.2
# google-api-core==2.8.2
# google-api-python-client==2.52.0
# google-auth==2.22.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.0
# google-cloud-core==2.3.1
# google-cloud-firestore==2.5.3
# google-cloud-storage==2.0.0
# google-crc32c==1.3.0
# google-resumable-media==2.3.3
# googleapis-common-protos==1.56.3
# grpcio==1.48.2
# grpcio-status==1.48.2
# hpack==3.0.0
# httplib2==0.22.0
# hyperframe==5.2.0
# hyperlink==19.0.0
# idna==2.8
# idna-ssl==1.1.0
# importlib-metadata==4.8.3
# importlib-resources==5.4.0
# incremental==22.10.0
# itemadapter==0.4.0
# itemloaders==1.0.4
# itsdangerous==2.0.1
# Jinja2==3.0.3
# jmespath==0.10.0
# joblib==0.14.1
# jsonpath-ng==1.4.3
# kafka==1.3.5
# lxml==4.3.4
# Markdown==3.3.7
# MarkupSafe==2.0.1
# msgpack==1.0.5
# multidict==4.7.4
# nltk==3.4.5
# numpy==1.19.5
# oauth2client==4.1.3
# oauthlib==3.2.2
# openpyxl==3.1.2
# pandas==1.1.5
# parsel==1.5.1
# pika==1.1.0
# Pillow==8.4.0
# ply==3.11
# Protego==0.2.1
# proto-plus==1.23.0
# protobuf==3.19.6
# pyasn1==0.4.5
# pyasn1-modules==0.2.5
# pycparser==2.19
# PyDispatcher==2.0.5
# PyHamcrest==1.9.0
# PyJWT==2.4.0
# pymongo==4.0.1
# pyOpenSSL==19.0.0
# pyparsing==3.1.1
# python-bsonjs==0.2.0
# python-dateutil==2.8.2
# pytz==2023.3.post1
# pytz-deprecation-shim==0.1.0.post0
# queuelib==1.5.0
# requests==2.22.0
# requests-file==1.4.3
# requests-oauthlib==2.0.0
# rsa==4.0
# scipy==1.3.1
# Scrapy==2.6.1
# scrapy-crawlera==1.6.0
# scrapy-rotating-proxies==0.6.2
# scrapy-selenium==0.0.7
# scrapy-splash==0.7.2
# Scrapy-UserAgents==0.0.1
# selenium-wire==1.0.11
# service-identity==18.1.0
# six==1.12.0
# soupsieve==2.3.2.post1
# sqlparse==0.4.4
# tk==0.1.0
# tldextract==2.2.2
# Twisted==22.4.0
# twisted-iocpsupport==1.0.2
# typing==3.7.4.1
# typing-extensions==3.7.4.1
# tzdata==2022.7
# tzlocal==2.0.0
# uritemplate==4.1.1
# urllib3==1.25.6
# w3lib==1.20.0
# webcolors==1.10
# Werkzeug==2.0.3
# xlrd==1.2.0
# yarl==1.4.2
# zipp==3.6.0
# zope.interface==4.6.0
# razorpay==1.4.2
# djangorestframework-simplejwt - 5.3.1
# psycopg2 - 2.9.9

asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0